Module frontier_template_node::rpc::eth

Structs

EthBlockDataCacheTask

Manage LRU caches for block data and their transaction statuses. These are large and take a lot of time to fetch from the database. Storing them in an LRU cache will allow to reduce database accesses when many subsequent requests are related to the same blocks.

EthDeps

Extra dependencies for Ethereum compatibility.

OverrideHandle

TxPool

Traits

EthConfig

StorageOverride

Something that can fetch Ethereum-related data. This trait is quite similar to the runtime API, and indeed oe implementation of it uses the runtime API. Having this trait is useful because it allows optimized implementations that fetch data from a State Backend with some assumptions about pallet-ethereum’s storage schema. Using such an optimized implementation avoids spawning a runtime and the overhead associated with it.

Functions

create_eth

Instantiate Ethereum-compatible RPC extensions.

overrides_handle

Type Definitions

FeeHistoryCache

FeeHistoryCacheLimit

Maximum fee history cache size.

FilterPool

On-memory stored filters created through the eth_newFilter RPC.