Module frontier_template_node::service

Service and ServiceFactory implementation. Specialized wrapper over substrate service.

Structs

Executor

MockTimestampInherentDataProvider

Provide a mock duration starting at 0 in millisecond for timestamp inherent. Each call will increment timestamp by slot_duration making Aura think time has passed.

NativeExecutor

A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

Constants

INHERENT_IDENTIFIER

TIMESTAMP

Functions

frontier_database_dir

new_full

Builds a new service for a full client.

new_light

Builds a new service for a light client.

new_partial

open_frontier_backend

remote_keystore

Type Definitions

ConsensusResult

FullBackend

FullClient

FullSelectChain